requirements.txt

fastapi
uvicorn
requests
mistralai
python-dotenv
